pyproject.toml
[project]
name = "chess-alphazero"
version = "0.1.0"
description = "AlphaZero-style chess training from scratch"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "python-chess>=1.999",
]
